Fixing the welcome page.
